Modify : 구조체별로 각기 다른 파일로 분류
